Add separate 'fetch' and 'merge' steps to VC API
This allows for the time-intensive part of package updates (i.e.
fetching) to be done without user interaction, and then merging can be
done while the user is present.

There is still a pain point when a package cannot be fetched, since
that will block further fetching in M-x straight-fetch-all until the
user says to continue. The eventual solution to this problem is
outlined in #103.

Merges #173.