Separate fetch and merge steps in the pulling operation.
The primary motivation is to be able to fetch all packages without
interrupting for user interaction (which takes network dependent time)
then in a separate step merge (which requires user interaction even if
only to hit the fast-forward button).

In a more conventional package manager these steps would be called
`update` and `upgrade` ;)

Depending on feedback the `pull` vc operations may be removed with the
high level pull-package and pull-all implemented in terms of fetch and
merge.