Fix modification checking for absolute :local-repo
Previously, if you specified a :local-repo that was an absolute path,
modifications would not be detected for that repository. This has now
been fixed.

:local-repo is now resolved to an absolute path using
`expand-file-name' during recipe conversion. This change results in a
one-time rebuild of all packages for the end user.

Performance may be improved, although this is untested, since the
find(1) command used for modification checking has been changed such
that it only searches in repositories for packages that are actually
being checked, rather than all repositories in (straight--repos-dir).