[#319] Merge pull request (#319)
[#318] Fix bug when cloning with :fork

The shell command to `git remote add fork` needs to happen in the
context of the package's repo, and previously this was happening in
`straight/repos`, which is usually the user's `.emacs.d` repo. The first
time it does this, it doesn't fail, since you can add the fork remote to
the repo; but the second time (for another recipe with a fork), it
finally fails because fork already exists as a remote. Both of these
executions are incorrect anyway, because it should be done in the cloned
repo of the package. Fortunately, fixing this is as easy as moving the
logic down, such that `default-directory` is set to `repo-dir`.