Improve lockfile UX (#375)
* New function straight-vc-commit-present-p: check if a SHA is fetched

This function should allow us to detect if a "git fetch" is necessary
upon thawing the repo.

* Automatically fetch the repo when the thawed-to commit is missing

This should address #58: If we're unable to find the commit locally,
the easiest way to get it is probably to fetch it.

* Normalize repos when thawing

Since we're already going to munge the repo state, we might as well
return it to the master branch and thus get it to a place where the
fetch can do something reasonable.

* Instead of decapitating via checkout, reset the HEAD ref

This should address #294: When thawing, we no longer leave detached
heads lying around, so freezing the repo will be much more likely to
be a high-touch operation. Also, this should leave the repos that we
*do* have in a much nicer state for people to work on.

* Remove note about running fetch before thawing.

With the auto-fetching behavior, I believe this should be unnecessary
now.

* Adjust straight-vc-check-out-commit to normalize the repo

This requires the interface for the function to change: Previously it
took just the local-repo, now it takes the whole recipe (because
that's what normalization uses).