Don't call pre-build functions unless building (#387)
Don't call the things on the `straight-use-package-pre-build-functions`
hook unless we're actually going to do a build.

I use the pre-build-functions hook to call `autoconf/configure/make`
for the vm package before the build happens and w/out this change it
was happening on every startup. With this change it only happens when
necessary.