Remove unnecessary hack
I figured out why this error was occurring. It was because I was
running the other Emacs version in a Docker container, and the user
happened to be root. Apparently, Docker will happily cause newly
created files to be written as root into your host filesystem, if they
are on a shared volume.