[Backport] Fix straight.el getting rebuilt twice
Because the timestamps saved in the build cache were not precise
enough, sometimes straight.elc would be compiled in the straight.el
directory and then straight.el the package would be symlinked and
compiled within the same second, thus causing the timestamp saved in
the build cache for straight.el to be the same as the file mtime on
straight.elc (or more precisely slightly newer, since the build cache
rounded the compilation timestamps down to the last second). This
triggered a second, spurious build of straight.el. Increasing build
cache timestamp precision to the microsecond level fixes the problem.