[Backport] Fix transactions not being finalized on error
The previous code relied on advices on exit-recursive-edit,
abort-recursive-edit, and top-level to catch returns to the top level
command loop. But that was obviously wishful thinking; in particular,
if you fail to clone a repository due to an SSH error, you'll return
to top level without invoking any of those functions, and then the
transaction will never be finalized, with the nasty effect that all
subsequent attempts to clone that package will be silently ignored (no
need to process a package more than once per transaction, right?).