Fix issue with CRLF line endings (#499)
insert-file-contents-literally prevents emacs from doing automatic line ending
detection and assumes that all line endings are \n. This is not true for
non-unix encodings such as windows (CRLF).

One way to fix this issue is try and make all the line ending regex more
verbose to deal with carriage returns. This is how this was handled in #416.
However I believe that the correct solution is to let Emacs automatically handle
the line endings (which is the default behavior). This avoids edge cases and
extra complexity in the code.