Add el-get recipe support
straight-recipes-el-get-list: Return a list of recipes available in
the `el-get` recipe repository.

straight-recipes-el-get-version: Return version of `el-get` retriever.

straight-recipes-el-get-retrieve: convert `el-get` recipes to straight.el
style recipes. Not all recipe keywords are supported.  `:prepare` and
`:post-init` are ignored as they may change the user's configuration.
`:autoloads` is partially supported. El-get allows a list of files as a
a value, which adds an extra autoloads file to load. This falls into the
same category as `:prepare` / `:post-init` and so it is ignored.

`el-get` recipes are stored in files, but may contain backquoted templates.
For example, many :build/* keyword values are intended to be dynamically
computed during installation. Recipes are returned with the whole list
backquoted, allowing `straight--convert-recipe` to handle evaluation of
backquote templates at the appropriate time. Otherwise we may
prematurely evaluate a template, or cache the result of such an
evaluation.