Allow multiple default branch names.
This makes it possible to deal with an ecosystem that may be diverging
away from the "master" branch naming. This change provides a different
default branch name, "main", and provides a list of alternate branch
names that can be tried.

In the various git operations, these alternate branches are tried
until one works.