Compute remote repository default branch names.
Use `straight-vc-git-default-branch' as a fallback.

The remote's default branch is calculated from its HEAD pointer.
The local name is computed via a series of fallbacks
(See the documentation of `straight-vc-git--default-branch-in-local-repo')

Remove `straight-vc-git-default-branch'.

We shouldn't need it anymore as we will be able to get a good local branch name.
The default remote branch will still be the default branch name - we
don't attempt to find a branch on remote that corresponds to the
branch on local.  That may be a future improvement, although it's not
clear what is the right behavior.

Remove `straight-vc-git--default-branch-in-local-repo'.

The logic this provided was just that the default branch was the
current branch, and in the bug we decided that wasn't the logic we
wanted. All calls have been migrated to
`straight-vc-git--default-remote-branch'.

When ensuring head without a remote ref, checkout default branch.

Also,fix the documentation string for `straight-vc-git--ensure-local'
because now, sometimes network communication happens if there's no
local reference indicating the default branch.

Rework ahead / behind logic when ensuring git head.

In `straight-vc-git--ensure-head', no longer consider whether we are
ahead or behind unless we're in the default branch.  With this, when
the branch is not the default branch, the user is asked to checkout,
merge, etc.