Simplify default branch logic.
This remove `straight-vc-git-default-branch'.  We should have no need
of it anymore, we always will be able to get a good local name.

The default remote branch will still be the default branch name - we
don't attempt to find a branch on remote that corresponds to the
branch on local.  That may be a future improvement, although it's not
clear what is the right behavior.