Clone with --no-single-branch to facilitate upstream development
When straight-vc-git-default-clone-depth is set to an integer, the
presence of the --depth option to git clone results in a fetch refspec
which is limited to the branch fetched:

    remote.origin.fetch=+refs/heads/master:refs/remotes/origin/master

(Incidentally, git clone does this even without the --single-branch
option which straight.el passes in this case.)

The (minor) problem with this is that if the user of straight.el wants
to contribute development of that package, and the package has a
workflow which involves other branches (e.g. Orgmode incorporates the
latest changes onto a maint branch), then a straight git fetch
origin (or equivalent via magit or some other UI) will not fetch any
other branches.

Arguably this is working as intended, but it does violate the
Principle of Least Surprise.  So simply change --single-branch to
--no-single-branch. This will make all upstream branches available,
but the clone depth is still respected so it's usually not a
significantly larger download, and the refspec is set up nicely for
fetching of all branches in the future:

    remote.origin.fetch=+refs/heads/*:refs/remotes/origin/*

Fixes #578