straight-bug-report: Improve report form formatting (#590)
* straight-bug-report: Improve report form formatting

`pp' and `pp-to-string' do a fair job of pretty printing the report
form, but they do not make optimal use of vertical space and tend to
produce wide forms. Multiple KEY VAL pairs are printed on the same line.
`straight-bug-report--report-form' returns a string which makes better
use of vertical space.