Org recipes: prevent duplicate byte-compilation/autoloads
Org's `oldorg` make target duplicates straight's byte-compilation and
autoload generation.
This duplicaiton is wasteful and error-prone (see: #624)
We're going with Org's `autoloads` make target because it generates
the necessary org-version.el file as well as autoloads.
Straight handles the byte-compilation.

Use "org" as build-dir name for org-plus-contrib
Otherwise we end up with both build/org-plus-contrib and build/org
when a package declares Org as a dependency after the user has installed
org-plus-contrib.