recipes-org-elpa-build: use tags when available (#851)
We accounted for a tag-less repo in #849.
However, the Org repo should usually have tags.
So now we use the original method (via git) and fall back to the
solution in #849 if necessary. This is analogous to what Org does.

Note there is a slight difference in the output of the two methods.
For example, the fallback method (which reads the "Version" metadata
field in org.el) will currently result in an ORGVERSION of "9.5",
while the git, tag-based version will result in "9.4.6".
I suspect the tag based version will be slightly more performant as
well.