Improve local-repo docs
From my initial reading of the docs (as a relative straight.el newbie) I
didn't realize that local-repo was meant for working with files on the
local file system, I assumed that it was something that you would use
with your own git repository server.

I hope that by being more specific about what a "local repository" is
that it will be easier for others to understand how to use the option.