Rename 'master' to 'main' (#1248)
I was holding off on this for a long time because it seemed like it was
going to break everything to change the default branch name, even though
I already did so years ago on all my other hundred repositories, but now
that I actually look at it, this should be a trivial change and
shouldn't affect anyone's existing configurations of `straight.el`. Feel
free to report any breakage and I can add further bridge code to assist
in avoiding impact.